requirements.txt
yapper-tts
coqui-tts
torch
